requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
sqlalchemy>=2.0
pyodbc>=5.1
python-dotenv>=1.0
passlib>=1.7
python-jose[cryptography]>=3.3
pydantic>=2.7
pandas>=2.2
